% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fsa.R
\name{plot_fsa}
\alias{plot_fsa}
\title{FSAPlot: plot_fsa}
\usage{
plot_fsa(fsa, dye, title = TRUE, x_limits)
}
\arguments{
\item{fsa}{A data.frame containing microsatellite fragment lengths and relative fluoresence produced using read_fsa.}

\item{dye}{Fluorescent dye bound to loci that one wishes to view. Only 6FAM, HEX, NED, or ROX are supported. This call can either be a single dye or a vector of multiple dyes. No default is provided.}

\item{title}{An argument defining whether the title of the fsa file is to be given at the head of a given plot. This argument must by TRUE or FALSE.}

\item{x_limits}{A vector of length 2 defining the basepair range to be viewed. Default is 100 to 500}
}
\description{
A function to plot fsa files produce from an ABI analyser and rendered into a data.frame using the read_fsa function. Plots are produced using ggplot2 by Hadley Wickham.
}
\details{
The default ladder for fragment length calibration is GS500. Note that only data.frames produced by read_fsa may be plotted with this function.
}
\examples{
source = "/home/fish332_microsat.fsa"
Peak_Dat = read_fsa(File = source, Lad_Chan = 4, Range = c(50:80))
plot_fsa(Peak_Dat, dye = c("6FAM", "NED"), title = TRUE, x_limits = c(50,80))
}
\keyword{ABI}
\keyword{fsa,}
\keyword{microsatellites,}
