% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fsa.r
\name{read_fsa}
\alias{read_fsa}
\title{FSAPlot: read_fsa}
\usage{
read_fsa(File, Lad_Chan = 4, Sig_Chan = 1:3, Range = c(100:500))
}
\arguments{
\item{File}{A path to a .fsa file to be opened.}

\item{Lad_Chan}{The loading channel in which the ladder may be found. Default is 4.}

\item{Sig_Chan}{Loading channels wherein sample reads lie. Default is 1 to 3.}

\item{Range}{Basepair range wherein peaks will be extracted. Default is 100 to 500.}
}
\description{
A function to read in .fsa files from an ABI analyser and produce a data.frame with relative fluoresences according to channel and basepair length.
}
\details{
The default ladder for fragment length calibration is GS500.
}
\examples{
source = "/home/fish332_microsat.fsa"
Peak_Dat = read_fsa(File = source, Lad_Chan = 4, Range = c(50:80))
}
\keyword{ABI}
\keyword{fsa,}
\keyword{microsatellites,}
